Update BOJ_13305.java
- 다맞음
- Int -> Long으로 변경

+++ AlgorithmStudy/src/acmicpc/BOJ_13305.java
@@ -14,20 +14,20 @@ public class BOJ_13305 {
 		BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 		
 		int N = Integer.parseInt(br.readLine());
-		int[] dis = new int[N-1];
-		int[] price = new int[N];
+		long[] dis = new long[N-1];
+		long[] price = new long[N];
 		
 		StringTokenizer st = new StringTokenizer(br.readLine());
-		for(int i = 0; i < N-1 ; i++) dis[i] = Integer.parseInt(st.nextToken());
+		for(int i = 0; i < N-1 ; i++) dis[i] = Long.parseLong(st.nextToken());
 		
 		st = new StringTokenizer(br.readLine());
-		for(int i = 0; i < N; i++) price[i] = Integer.parseInt(st.nextToken());
+		for(int i = 0; i < N; i++) price[i] = Long.parseLong(st.nextToken());
 		
 		for(int i = 1; i < N; i++) {
 			if(price[i-1] < price[i]) price[i] = price[i-1];
 		}
 		
-		int res = 0;
+		long res = 0;
 		for(int i = 0; i < N-1; i++) {
 			res += dis[i] * price[i];
 		}
